+++ pehlishuruaat/app/src/main/java/com/sachinkrdkflare/pehlishuruaat/MainActivity.java
@@ -1,6 +1,7 @@
 package com.sachinkrdkflare.pehlishuruaat;
 
-import android.graphics.Typeface;
+import android.content.Intent;
+import android.os.Handler;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -12,110 +13,24 @@ import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
 
-    private ViewPager mSlideViewPager;
-    private LinearLayout mDotLayout;
-
-    private TextView[] mDots;
-
-    private SliderAdapter sliderAdapter;
-
-    private Button mNextBtn;
-    private Button mBackBtn;
-
-    private int mCurrentPage;
-
-
+    private static int SPLASH_TIME_OUT = 2000;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mSlideViewPager = (ViewPager) findViewById(R.id.slideViewPager);
-        mDotLayout = (LinearLayout) findViewById(R.id.dotslayout);
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
 
-        mNextBtn = (Button) findViewById(R.id.nextBtn);
-        mBackBtn = (Button) findViewById(R.id.prevBtn);
-
-        sliderAdapter = new SliderAdapter(this);
-
-        mSlideViewPager.setAdapter(sliderAdapter);
-
-        addDotsIndicator(0);
-
-        mSlideViewPager.addOnPageChangeListener(viewListener);
+                Intent homeIntent = new Intent(MainActivity.this, Walkthrough.class);
+                startActivity(homeIntent);
+                finish();
+            }
+        }, SPLASH_TIME_OUT);
     }
 
-    public void addDotsIndicator(int position){
-
-        mDots = new TextView[3];
-        mDotLayout.removeAllViews();
-
-        for (int i = 0; i < mDots.length; i++){
-
-            mDots[i] = new TextView(this);
-            mDots[i].setText(Html.fromHtml("&#8226"));
-            mDots[i].setTextSize(35);
-            mDots[i].setTextColor(getResources().getColor(R.color.colorTransparentWhite));
-
-            mDotLayout.addView(mDots[i]);
-
-        }
-
-        if(mDots.length > 0){
-
-            mDots[position].setTextColor(getResources().getColor(R.color.colorTransparentWhite));
-        }
-    }
-
-    ViewPager.OnPageChangeListener viewListener;
-
-    {
-        viewListener = new ViewPager.OnPageChangeListener() {
-            @Override
-            public void onPageScrolled(int i, float v, int i1) {
-
-            }
-
-            @Override
-            public void onPageSelected(int i) {
-
-                addDotsIndicator(i);
-                mCurrentPage = i;
-
-                if (i == 0) {
-
-                    mNextBtn.setEnabled(true);
-                    mBackBtn.setEnabled(false);
-                    mBackBtn.setVisibility(View.INVISIBLE);
-
-                    mNextBtn.setText("Next");
-                    mBackBtn.setText("");
-                } else if (i == mDots.length - 1) {
-
-                    mNextBtn.setEnabled(true);
-                    mBackBtn.setEnabled(true);
-                    mBackBtn.setVisibility(View.VISIBLE);
-
-                    mNextBtn.setText("Finish");
-                    mBackBtn.setText("Back");
-                } else {
-
-                    mNextBtn.setEnabled(true);
-                    mBackBtn.setEnabled(true);
-                    mBackBtn.setVisibility(View.VISIBLE);
-
-                    mNextBtn.setText("Next");
-                    mBackBtn.setText("Back");
-                }
-
-            }
-
-            @Override
-            public void onPageScrollStateChanged(int i) {
-
-            }
-        };
-    }
 }
