+++ pehlishuruaat/app/src/main/res/layout/activity_main.xml
@@ -7,49 +7,49 @@
     android:background="@color/colorTransparentWhite"
     tools:context=".MainActivity">
 
-    <android.support.v4.view.ViewPager
-        android:id="@+id/slideViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="659dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginLeft="0dp"
-        android:layout_marginTop="0dp"></android.support.v4.view.ViewPager>
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="250dp"
+        android:layout_height="250dp"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="100dp"
+        app:srcCompat="@drawable/donate_blood" />
 
-    <LinearLayout
-        android:id="@+id/dotslayout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="220dp"
+        android:layout_height="50dp"
+        android:layout_marginTop="350dp"
+        android:layout_marginLeft="100dp"
+        android:layout_centerHorizontal="true"
+        android:fontFamily="@font/apple_gothic_neo"
+        android:text="Blood Donation"
+        android:textColor="@color/colorRed"
+        android:textSize="36sp"
+        android:textStyle="bold"
+        android:visibility="visible" />
+
+
+    <TextView
+        android:id="@+id/textView3"
+        android:layout_width="200dp"
+        android:layout_height="40dp"
+        android:layout_below="@+id/imageView"
+        android:layout_marginTop="55dp"
+        android:layout_centerHorizontal="true"
+        android:fontFamily="@font/avari_light"
+        android:text="Every life is precious"
+        android:textSize="21sp"
+        android:visibility="visible" />
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="120dp"
+        android:layout_height="60dp"
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
-        android:layout_marginBottom="30dp"
-        android:orientation="horizontal"
-        android:padding="20dp"></LinearLayout>
-
-    <Button
-        android:id="@+id/prevBtn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="10dp"
-        android:layout_marginTop="655dp"
-        android:layout_marginRight="0dp"
-        android:background="@android:color/transparent"
-        android:text="BACK"
-        android:textColor="@color/colorTransparentWhite" />
-
-    <Button
-        android:id="@+id/nextBtn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="310dp"
-        android:layout_marginTop="655dp"
-        android:layout_marginRight="0dp"
-        android:background="@android:color/transparent"
-        android:text="NEXT"
-        android:textColor="@color/colorTransparentWhite"
-        android:textSize="18sp" />
+        app:srcCompat="@drawable/verved" />
 
 
 </RelativeLayout>
