+++ pehlishuruaat/app/src/main/res/values/colors.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#262938</color>
-    <color name="colorPrimaryDark">#181A24</color>
+    <color name="colorPrimary">#E5E8E8</color>
+    <color name="colorPrimaryDark">#CACFD2</color>
     <color name="colorAccent">#FF4001</color>
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorTransparentWhite">#FFFFFF</color>
+    <color name="colorRed">#FF0000</color>
 </resources>
